requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlite3

# AI/Agent Dependencies
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.5
langsmith==0.0.69

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gmail-api==0.0.4

# Web Scraping & HTTP
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
httpx==0.25.2

# Geospatial & Maps
geopy==2.4.0
shapely==2.0.2
geopandas==0.14.1

# Utility
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0 